+++ modules/app/src/test/scala/cards/nine/app/ui/wizard/jobs/NewConfigurationJobsSpecification.scala
@@ -1,11 +1,10 @@
 package cards.nine.app.ui.wizard.jobs
 
 import cards.nine.app.di.Injector
-import cards.nine.app.ui.wizard.jobs.uiactions.{NewConfigurationUiActions, VisibilityUiActions}
+import cards.nine.app.ui.wizard.jobs.uiactions.{NewConfigurationUiActions, VisibilityUiActions, WizardUiActions}
 import cards.nine.commons.contexts.ContextSupport
 import cards.nine.commons.test.TaskServiceSpecification
 import cards.nine.commons.test.data.{ApiTestData, ApplicationTestData, MomentTestData}
-import cards.nine.models.{MomentData, MomentTimeSlot}
 import cards.nine.models.types._
 import cards.nine.process.collection.{CollectionException, CollectionProcess}
 import cards.nine.process.device.{AppException, DeviceException, DeviceProcess}
@@ -52,6 +51,8 @@ trait NewConfigurationJobsSpecification
     newConfigurationUiActions.loadFifthStep() returns serviceRight(Unit)
     newConfigurationUiActions.loadSixthStep() returns serviceRight(Unit)
 
+    val wizardUiActions = mock[WizardUiActions]
+
     val mockInjector: Injector = mock[Injector]
 
     val mockDeviceProcess = mock[DeviceProcess]
@@ -74,7 +75,7 @@ trait NewConfigurationJobsSpecification
 
     mockInjector.widgetsProcess returns mockWidgetsProcess
 
-    val newConfigurationJobs = new NewConfigurationJobs(newConfigurationUiActions, visibilityUiActions)(contextWrapper) {
+    val newConfigurationJobs = new NewConfigurationJobs(wizardUiActions, newConfigurationUiActions, visibilityUiActions)(contextWrapper) {
 
       override lazy val di: Injector = mockInjector
 
@@ -96,9 +97,9 @@ class NewConfigurationJobsSpec
       mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
       mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory)
 
-      newConfigurationJobs.loadBetterCollections(hidePrevious = true) mustRightUnit
+      newConfigurationJobs.loadBetterCollections().mustRightUnit
 
-      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(any)
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(true)
       there was one(mockDeviceProcess).resetSavedItems()
       there was one(mockDeviceProcess).synchronizeInstalledApps(any)
     }
@@ -109,9 +110,9 @@ class NewConfigurationJobsSpec
       mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
       mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory map (_.copy(category = Misc)))
 
-      newConfigurationJobs.loadBetterCollections(hidePrevious = true) mustRightUnit
+      newConfigurationJobs.loadBetterCollections().mustRightUnit
 
-      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(any)
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(true)
       there was one(mockDeviceProcess).resetSavedItems()
       there was one(mockDeviceProcess).synchronizeInstalledApps(any)
     }
@@ -122,9 +123,9 @@ class NewConfigurationJobsSpec
       mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
       mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory map (_.copy(packages = Seq.empty)))
 
-      newConfigurationJobs.loadBetterCollections(hidePrevious = true) mustRightUnit
+      newConfigurationJobs.loadBetterCollections().mustRightUnit
 
-      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(any)
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(true)
       there was one(mockDeviceProcess).resetSavedItems()
       there was one(mockDeviceProcess).synchronizeInstalledApps(any)
     }
@@ -135,12 +136,73 @@ class NewConfigurationJobsSpec
       mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
       mockCollectionProcess.rankApps()(any) returns serviceLeft(collectionException)
 
-      newConfigurationJobs.loadBetterCollections(hidePrevious = true).mustLeft[CollectionException]
+      newConfigurationJobs.loadBetterCollections().mustLeft[CollectionException]
 
-      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(any)
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(true)
       there was one(mockDeviceProcess).resetSavedItems()
       there was one(mockDeviceProcess).synchronizeInstalledApps(any)
     }
+  }
+
+  "rollbackLoadBetterCollections" should {
+
+    "return a Seq of PackagesByCategory and clean collections" in new NewConfigurationJobsScope {
+
+      mockDeviceProcess.resetSavedItems() returns serviceRight(Unit)
+      mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
+      mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory)
+      mockCollectionProcess.cleanCollections() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackLoadBetterCollections().mustRightUnit
+
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(false)
+      there was one(mockDeviceProcess).resetSavedItems()
+      there was one(mockDeviceProcess).synchronizeInstalledApps(any)
+      there was one(mockCollectionProcess).cleanCollections()
+    }
+
+    "return a Seq empty if the category of the collections are Misc and clean collections" in new NewConfigurationJobsScope {
+
+      mockDeviceProcess.resetSavedItems() returns serviceRight(Unit)
+      mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
+      mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory map (_.copy(category = Misc)))
+      mockCollectionProcess.cleanCollections() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackLoadBetterCollections().mustRightUnit
+
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(false)
+      there was one(mockDeviceProcess).resetSavedItems()
+      there was one(mockDeviceProcess).synchronizeInstalledApps(any)
+    }
+
+    "return a Seq empty if the packages size of the collections are less than 3 and clean collections" in new NewConfigurationJobsScope {
+
+      mockDeviceProcess.resetSavedItems() returns serviceRight(Unit)
+      mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
+      mockCollectionProcess.rankApps()(any) returns serviceRight(seqPackagesByCategory map (_.copy(packages = Seq.empty)))
+      mockCollectionProcess.cleanCollections() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackLoadBetterCollections().mustRightUnit
+
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(false)
+      there was one(mockDeviceProcess).resetSavedItems()
+      there was one(mockDeviceProcess).synchronizeInstalledApps(any)
+    }
+
+    "return a CollectionException when the service returns an exception" in new NewConfigurationJobsScope {
+
+      mockDeviceProcess.resetSavedItems() returns serviceRight(Unit)
+      mockDeviceProcess.synchronizeInstalledApps(any) returns serviceRight(Unit)
+      mockCollectionProcess.rankApps()(any) returns serviceLeft(collectionException)
+      mockCollectionProcess.cleanCollections() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackLoadBetterCollections().mustLeft[CollectionException]
+
+      there was one(visibilityUiActions).hideFistStepAndShowLoadingBetterCollections(false)
+      there was one(mockDeviceProcess).resetSavedItems()
+      there was one(mockDeviceProcess).synchronizeInstalledApps(any)
+    }
+
   }
 
   "saveCollections" should {
@@ -183,12 +245,12 @@ class NewConfigurationJobsSpec
 
   "loadMomentWithWifi" should {
 
-    "return a Seq that contains all configured networks" in new NewConfigurationJobsScope {
+    "return a valid response when getConfiguredNetworks return valid sequence" in new NewConfigurationJobsScope {
 
       val networks = 0 to 10 map (c => s"Networks $c")
       mockDeviceProcess.getConfiguredNetworks(any) returns serviceRight(networks)
 
-      newConfigurationJobs.loadMomentWithWifi(hidePrevious = true).mustRightUnit
+      newConfigurationJobs.loadMomentWithWifi().mustRightUnit
 
       there was one(visibilityUiActions).hideThirdStep()
     }
@@ -197,9 +259,46 @@ class NewConfigurationJobsSpec
 
       mockDeviceProcess.getConfiguredNetworks(any) returns serviceLeft(deviceException)
 
-      newConfigurationJobs.loadMomentWithWifi(hidePrevious = true).mustLeft[DeviceException]
+      newConfigurationJobs.loadMomentWithWifi().mustLeft[DeviceException]
 
       there was one(visibilityUiActions).hideThirdStep()
+    }
+
+  }
+
+  "rollbackMomentWithWifi" should {
+
+    "return a valid response and delete moments and widgets when rollback" in new NewConfigurationJobsScope {
+
+      val networks = 0 to 10 map (c => s"Networks $c")
+      mockDeviceProcess.getConfiguredNetworks(any) returns serviceRight(networks)
+
+      wizardUiActions.showErrorGeneral() returns serviceRight(Unit)
+      visibilityUiActions.cleanNewConfiguration() returns serviceRight(Unit)
+
+      mockMomentProcess.deleteAllMoments() returns serviceRight(Unit)
+      mockWidgetsProcess.deleteAllWidgets() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackMomentWithWifi().mustRightUnit
+
+      there was one(wizardUiActions).showErrorGeneral()
+      there was one(visibilityUiActions).cleanNewConfiguration()
+      there was one(mockMomentProcess).deleteAllMoments()
+      there was one(mockWidgetsProcess).deleteAllWidgets()
+    }
+
+    "return a DeviceException when the service returns an exception" in new NewConfigurationJobsScope {
+
+      mockDeviceProcess.getConfiguredNetworks(any) returns serviceLeft(deviceException)
+
+      wizardUiActions.showErrorGeneral() returns serviceRight(Unit)
+      visibilityUiActions.cleanNewConfiguration() returns serviceRight(Unit)
+
+      mockMomentProcess.deleteAllMoments() returns serviceRight(Unit)
+      mockWidgetsProcess.deleteAllWidgets() returns serviceRight(Unit)
+
+      newConfigurationJobs.rollbackMomentWithWifi().mustLeft[DeviceException]
+
     }
 
   }